* commit '24d63a838211db019bf7ea970a0d844d8033b729':
  AArch64: Use long for pointers in media classes

+++ media/java/android/media/RemoteDisplay.java
@@ -38,12 +38,12 @@ public final class RemoteDisplay {
     private final Listener mListener;
     private final Handler mHandler;
 
-    private int mPtr;
+    private long mPtr;
 
-    private native int nativeListen(String iface);
-    private native void nativeDispose(int ptr);
-    private native void nativePause(int ptr);
-    private native void nativeResume(int ptr);
+    private native long nativeListen(String iface);
+    private native void nativeDispose(long ptr);
+    private native void nativePause(long ptr);
+    private native void nativeResume(long ptr);
 
     private RemoteDisplay(Listener listener, Handler handler) {
         mListener = listener;
